+++ src/core/array.spec.ts
@@ -1,4 +1,6 @@
+import { isNull, isNumber } from '../utils'
 import { VOArray, VOArrayOptions } from './array'
+import { VOError } from './errors'
 
 describe('VOArray', () => {
   it('Should return a class that can be extended', () => {
@@ -60,7 +62,7 @@ describe('VOArray', () => {
 
     class Base {
       constructor(error: boolean) {
-        if (error) throw Error('Test error')
+        if (error) throw new VOError('Test error')
       }
       valueOf(): 456 {
         return 456
@@ -73,8 +75,8 @@ describe('VOArray', () => {
     expect(fn).toThrowMatching(
       (errArray): boolean =>
         Array.isArray(errArray) &&
-        errArray.every(err => err instanceof Error) &&
-        errArray.every(err => typeof err.index === 'number'),
+        errArray.every(VOError.is) &&
+        errArray.every(err => isNumber(err.path.toArray().pop())),
     )
   })
 
@@ -109,7 +111,7 @@ describe('VOArray', () => {
 
     for (const test of tests) {
       const fn = () => VOArray(Base, test)
-      if (test.error === null) expect(fn).not.toThrow()
+      if (isNull(test.error)) expect(fn).not.toThrow()
       else expect(fn).toThrowError(test.error)
     }
   })
@@ -156,7 +158,7 @@ describe('VOArray', () => {
 
     class Base {
       constructor(raw: number) {
-        throw Error('Test error')
+        throw new VOError('Test error')
       }
       valueOf(): 456 {
         return 456
@@ -171,9 +173,8 @@ describe('VOArray', () => {
         (errArray): boolean =>
           Array.isArray(errArray) &&
           errArray.length === maxErrors &&
-          errArray.every(err => err instanceof Error) &&
-          errArray.every(err => typeof err.index === 'number') &&
-          errArray.every((err, i) => err.index === i),
+          errArray.every(VOError.is) &&
+          errArray.every((err, i) => err.path.toArray().pop() === i),
       )
     }
   })