+++ src/utils/index.ts
@@ -1,1 +1,3 @@
 export * from './either'
+export * from './guards'
+export * from './object-path'