+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Generate a website for the API docs
 - Use Docker for the website
 
+### Changed
+
+- #5 Errors show path to property
+
 ## [1.2.1](https://github.com/LucasPaganini/value-objects/releases/tag/v1.2.1) - 2020-06-08
 
 ### Added